Make /log dynamic

+++ src/console.c
@@ -7,11 +7,79 @@
 
 extern Editor_State console_view_es;
 extern Editor_State console_input_es;
-extern Editor_State main_text_es[MAX_FILES_OPEN];
+
+bool console_keep_log = false;
 
 void update_console()
 {
+  if (console_keep_log)
+  {
+    Editor_State* es = ui_get_focused_editor();
+    s64 cursor_offset = 0;
+    u8 aux_str[64];
+    s32 n;
 
+    if (get_tid_text_size(console_view_es.main_buffer_tid) > 0)
+    delete_text(console_view_es.main_buffer_tid, null, get_tid_text_size(console_view_es.main_buffer_tid), 0);
+
+    insert_text(console_view_es.main_buffer_tid, "HoEXditor Console", sizeof("HoEXditor Console") - 1, cursor_offset);
+    cursor_offset += sizeof("HoEXditor Console") - 1;
+
+    insert_text(console_view_es.main_buffer_tid, "\n\nCursor offset: ", sizeof("\n\nCursor offset: ") - 1, cursor_offset);
+    cursor_offset += sizeof("\n\nCursor offset: ") - 1;
+    n = s64_to_str_base10(es->cursor_info.cursor_offset, aux_str);
+    insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
+    cursor_offset += n;
+
+    insert_text(console_view_es.main_buffer_tid, "\nNext line count: ", sizeof("\nNext line count: ") - 1, cursor_offset);
+    cursor_offset += sizeof("\nNext line count: ") - 1;
+    n = s64_to_str_base10(es->cursor_info.next_line_count, aux_str);
+    insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
+    cursor_offset += n;
+
+    insert_text(console_view_es.main_buffer_tid, "\nPrev line count: ", sizeof("\nPrev line count: ") - 1, cursor_offset);
+    cursor_offset += sizeof("\nPrev line count: ") - 1;
+    n = s64_to_str_base10(es->cursor_info.previous_line_count, aux_str);
+    insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
+    cursor_offset += n;
+
+    insert_text(console_view_es.main_buffer_tid, "\nSnap cursor column: ", sizeof("\nSnap cursor column: ") - 1, cursor_offset);
+    cursor_offset += sizeof("\nSnap cursor column: ") - 1;
+    n = s64_to_str_base10(es->cursor_info.cursor_snaped_column, aux_str);
+    insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
+    cursor_offset += n;
+
+    insert_text(console_view_es.main_buffer_tid, "\nCursor column: ", sizeof("\nCursor column: ") - 1, cursor_offset);
+    cursor_offset += sizeof("\nCursor column: ") - 1;
+    n = s64_to_str_base10(es->cursor_info.cursor_column, aux_str);
+    insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
+    cursor_offset += n;
+
+    insert_text(console_view_es.main_buffer_tid, "\nCursor line: ", sizeof("\nCursor line: ") - 1, cursor_offset);
+    cursor_offset += sizeof("\nCursor line: ") - 1;
+    cursor_info cinfo = get_cursor_info(es->main_buffer_tid, es->cursor_info.cursor_offset);
+    n = s64_to_str_base10(cinfo.line_number.lf, aux_str);
+    insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
+    cursor_offset += n;
+
+    insert_text(console_view_es.main_buffer_tid, "\nText Size: ", sizeof("\nText Size: ") - 1, cursor_offset);
+    cursor_offset += sizeof("\nText Size: ") - 1;
+    n = s64_to_str_base10(get_tid_text_size(es->main_buffer_tid), aux_str);
+    insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
+    cursor_offset += n;
+
+    insert_text(console_view_es.main_buffer_tid, "\nBuffer Valid Bytes: ", sizeof("\nBuffer Valid Bytes: ") - 1, cursor_offset);
+    cursor_offset += sizeof("\nBuffer Valid Bytes: ") - 1;
+    n = s64_to_str_base10(get_tid_valid_bytes(es->main_buffer_tid), aux_str);
+    insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
+    cursor_offset += n;
+
+    insert_text(console_view_es.main_buffer_tid, "\nLast Line: ", sizeof("\nLast Line: ") - 1, cursor_offset);
+    cursor_offset += sizeof("\nLast Line: ") - 1;
+    n = s64_to_str_base10(es->cursor_info.last_line, aux_str);
+    insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
+    cursor_offset += n;
+  }
 }
 
 s32 console_char_handler(s32 key)
@@ -160,6 +228,7 @@ void run_help_command(s32 argc, u8* argv[])
 
 void run_open_command(s32 argc, u8* argv[])
 {
+  Editor_State* es = ui_get_focused_editor();
   u8 success_text[] = "File opened successfully";
   s32 success_text_size = sizeof("File opened successfully") - 1;
   u8 unknown_error_text[] = "Error: Unknown Error.";
@@ -178,7 +247,7 @@ void run_open_command(s32 argc, u8* argv[])
     {
       if (!ui_open_file(false, argv[1]))
       {
-        update_buffer(&main_text_es[0]);
+        update_buffer(es);
         insert_text(console_view_es.main_buffer_tid, success_text, success_text_size, 0);
       }
       else
@@ -193,74 +262,24 @@ void run_open_command(s32 argc, u8* argv[])
 
 void run_log_command(s32 argc, u8* argv[])
 {
-  s64 cursor_offset = 0;
-  u8 aux_str[64];
-  s32 n;
+  console_keep_log = !console_keep_log;
 
-  if (get_tid_text_size(console_view_es.main_buffer_tid) > 0)
-  delete_text(console_view_es.main_buffer_tid, null, get_tid_text_size(console_view_es.main_buffer_tid), 0);
+  // if log was turned off
+  if (!console_keep_log)
+  {
+    u8 log_turned_off_text[] = "/log turned off.";
+    s32 log_turned_off_text_size = sizeof("/log turned off.") - 1;
 
-  insert_text(console_view_es.main_buffer_tid, "HoEXditor Console", sizeof("HoEXditor Console") - 1, cursor_offset);
-  cursor_offset += sizeof("HoEXditor Console") - 1;
+    if (get_tid_text_size(console_view_es.main_buffer_tid) > 0)
+      delete_text(console_view_es.main_buffer_tid, null, get_tid_text_size(console_view_es.main_buffer_tid), 0);
 
-  insert_text(console_view_es.main_buffer_tid, "\n\nCursor offset: ", sizeof("\n\nCursor offset: ") - 1, cursor_offset);
-  cursor_offset += sizeof("\n\nCursor offset: ") - 1;
-  n = s64_to_str_base10(main_text_es[0].cursor_info.cursor_offset, aux_str);
-  insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
-  cursor_offset += n;
-
-  insert_text(console_view_es.main_buffer_tid, "\nNext line count: ", sizeof("\nNext line count: ") - 1, cursor_offset);
-  cursor_offset += sizeof("\nNext line count: ") - 1;
-  n = s64_to_str_base10(main_text_es[0].cursor_info.next_line_count, aux_str);
-  insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
-  cursor_offset += n;
-
-  insert_text(console_view_es.main_buffer_tid, "\nPrev line count: ", sizeof("\nPrev line count: ") - 1, cursor_offset);
-  cursor_offset += sizeof("\nPrev line count: ") - 1;
-  n = s64_to_str_base10(main_text_es[0].cursor_info.previous_line_count, aux_str);
-  insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
-  cursor_offset += n;
-
-  insert_text(console_view_es.main_buffer_tid, "\nSnap cursor column: ", sizeof("\nSnap cursor column: ") - 1, cursor_offset);
-  cursor_offset += sizeof("\nSnap cursor column: ") - 1;
-  n = s64_to_str_base10(main_text_es[0].cursor_info.cursor_snaped_column, aux_str);
-  insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
-  cursor_offset += n;
-
-  insert_text(console_view_es.main_buffer_tid, "\nCursor column: ", sizeof("\nCursor column: ") - 1, cursor_offset);
-  cursor_offset += sizeof("\nCursor column: ") - 1;
-  n = s64_to_str_base10(main_text_es[0].cursor_info.cursor_column, aux_str);
-  insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
-  cursor_offset += n;
-
-  insert_text(console_view_es.main_buffer_tid, "\nCursor line: ", sizeof("\nCursor line: ") - 1, cursor_offset);
-  cursor_offset += sizeof("\nCursor line: ") - 1;
-  cursor_info cinfo = get_cursor_info(main_text_es[0].main_buffer_tid, main_text_es[0].cursor_info.cursor_offset);
-  n = s64_to_str_base10(cinfo.line_number.lf, aux_str);
-  insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
-  cursor_offset += n;
-
-  insert_text(console_view_es.main_buffer_tid, "\nText Size: ", sizeof("\nText Size: ") - 1, cursor_offset);
-  cursor_offset += sizeof("\nText Size: ") - 1;
-  n = s64_to_str_base10(get_tid_text_size(main_text_es[0].main_buffer_tid), aux_str);
-  insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
-  cursor_offset += n;
-
-  insert_text(console_view_es.main_buffer_tid, "\nBuffer Valid Bytes: ", sizeof("\nBuffer Valid Bytes: ") - 1, cursor_offset);
-  cursor_offset += sizeof("\nBuffer Valid Bytes: ") - 1;
-  n = s64_to_str_base10(get_tid_valid_bytes(main_text_es[0].main_buffer_tid), aux_str);
-  insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
-  cursor_offset += n;
-
-  insert_text(console_view_es.main_buffer_tid, "\nLast Line: ", sizeof("\nLast Line: ") - 1, cursor_offset);
-  cursor_offset += sizeof("\nLast Line: ") - 1;
-  n = s64_to_str_base10(main_text_es[0].cursor_info.last_line, aux_str);
-  insert_text(console_view_es.main_buffer_tid, aux_str, n, cursor_offset);
-  cursor_offset += n;
+    insert_text(console_view_es.main_buffer_tid, log_turned_off_text, log_turned_off_text_size, 0);
+  }
 }
 
 void run_save_command(s32 argc, u8* argv[])
 {
+  Editor_State* es = ui_get_focused_editor();
   u8 success_text[] = "File saved successfully";
   s32 success_text_size = sizeof("File saved successfully") - 1;
   u8 unknown_error_text[] = "Error: Unknown Error.";
@@ -279,7 +298,7 @@ void run_save_command(s32 argc, u8* argv[])
 
     if (does_path_exist(path))
     {
-      if (!save_file(main_text_es[0].main_buffer_tid, argv[1]))
+      if (!save_file(es->main_buffer_tid, argv[1]))
         insert_text(console_view_es.main_buffer_tid, success_text, success_text_size, 0);
       else
         insert_text(console_view_es.main_buffer_tid, unknown_error_text, unknown_error_text_size, 0);
@@ -295,6 +314,7 @@ void run_save_command(s32 argc, u8* argv[])
 
 void run_new_empty_file_command(s32 argc, u8* argv[])
 {
+  Editor_State* es = ui_get_focused_editor();
   u8 success_text[] = "Empty file created successfully.";
   s32 success_text_size = sizeof("Empty file opened successfully") - 1;
   u8 unknown_error_text[] = "Error: Unknown Error.";
@@ -305,7 +325,7 @@ void run_new_empty_file_command(s32 argc, u8* argv[])
 
   if (!ui_open_file(true, null))
   {
-    update_buffer(&main_text_es[0]);
+    update_buffer(es);
     insert_text(console_view_es.main_buffer_tid, success_text, success_text_size, 0);
   }
   else
